공지 목록조회 페이징 작성중

+++ src/main/java/com/fooddiary/api/repository/NoticeDynamicConditionRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.domain.Pageable;
 import com.fooddiary.api.entity.notice.Notice;
 
 public interface NoticeDynamicConditionRepository {
-    int selectCount(String title, String content, Boolean available, LocalDate noticeAt);
+    long selectCount(String title, String content, Boolean available, LocalDate noticeAt);
 
     List<Notice> selectList(String title, String content, Boolean available, LocalDate noticeAt,
                                    Pageable pageable);
